remote-communicationのcommunicationのアダプター構成をリファクタリング
