+++ remote-communication/pkg/communication/adapters/rabbitmq/mock_data_test.go
@@ -3,6 +3,7 @@ package rabbitmq
 import (
 	"context"
 	"remote-communication/pkg/communication/domain/communication"
+	"remote-communication/pkg/communication/service"
 
 	crm "github.com/Tomofiles/skysign_cloud_v2/skysign-common/pkg/common/adapters/rabbitmq"
 
@@ -10,6 +11,27 @@ import (
 )
 
 const DefaultCommunicationID = communication.ID("communication-id")
+
+type manageCommunicationServiceMock struct {
+	mock.Mock
+	ID string
+}
+
+func (s *manageCommunicationServiceMock) CreateCommunication(
+	command service.CreateCommunicationCommand,
+) error {
+	ret := s.Called()
+	s.ID = command.GetID()
+	return ret.Error(0)
+}
+
+func (s *manageCommunicationServiceMock) DeleteCommunication(
+	command service.DeleteCommunicationCommand,
+) error {
+	ret := s.Called()
+	s.ID = command.GetID()
+	return ret.Error(0)
+}
 
 type channelMockPublish struct {
 	mock.Mock
@@ -42,11 +64,51 @@ func (ch *channelMockPublish) Close() error {
 	panic("implement me")
 }
 
-type publishHandlerMock struct {
+type pubSubManagerMock struct {
 	publishHandlers []func(ch crm.Channel, e interface{})
 }
 
-func (h *publishHandlerMock) SetPublishHandler(handler func(ch crm.Channel, e interface{})) error {
+func (h *pubSubManagerMock) SetPublishHandler(handler func(ch crm.Channel, e interface{})) error {
 	h.publishHandlers = append(h.publishHandlers, handler)
+	return nil
+}
+
+type publishHandlerMock struct {
+	consumers []consumer
+}
+
+func (h *publishHandlerMock) SetConsumer(ctx context.Context, exchangeName, queueName string, handler func([]byte)) error {
+	h.consumers = append(
+		h.consumers,
+		consumer{
+			exchangeName: exchangeName,
+			queueName:    queueName,
+			handler:      handler,
+		})
+	return nil
+}
+
+type consumer struct {
+	exchangeName, queueName string
+	handler                 func([]byte)
+}
+
+type eventHandlerMock struct {
+	events1, events2 []byte
+}
+
+func (h *eventHandlerMock) HandleCommunicationIDGaveEvent(
+	ctx context.Context,
+	event []byte,
+) error {
+	h.events1 = append(h.events1, event...)
+	return nil
+}
+
+func (h *eventHandlerMock) HandleCommunicationIDRemovedEvent(
+	ctx context.Context,
+	event []byte,
+) error {
+	h.events2 = append(h.events2, event...)
 	return nil
 }
