+++ remote-communication/pkg/communication/adapters/rabbitmq/subscribe_event_handler.go
@@ -1,4 +1,4 @@
-package ports
+package rabbitmq
 
 import (
 	"context"
