+++ remote-communication/pkg/communication/adapters/rabbitmq/subscribe_event_handler_test.go
@@ -1,7 +1,6 @@
-package ports
+package rabbitmq
 
 import (
-	"context"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -53,44 +52,4 @@ func TestSubscribeEventHandlerCommunicationIdRemovedEvent(t *testing.T) {
 
 	a.Empty(evt.events1)
 	a.Equal(evt.events2, EventByte)
-}
-
-type publishHandlerMock struct {
-	consumers []consumer
-}
-
-func (h *publishHandlerMock) SetConsumer(ctx context.Context, exchangeName, queueName string, handler func([]byte)) error {
-	h.consumers = append(
-		h.consumers,
-		consumer{
-			exchangeName: exchangeName,
-			queueName:    queueName,
-			handler:      handler,
-		})
-	return nil
-}
-
-type consumer struct {
-	exchangeName, queueName string
-	handler                 func([]byte)
-}
-
-type eventHandlerMock struct {
-	events1, events2 []byte
-}
-
-func (h *eventHandlerMock) HandleCommunicationIDGaveEvent(
-	ctx context.Context,
-	event []byte,
-) error {
-	h.events1 = append(h.events1, event...)
-	return nil
-}
-
-func (h *eventHandlerMock) HandleCommunicationIDRemovedEvent(
-	ctx context.Context,
-	event []byte,
-) error {
-	h.events2 = append(h.events2, event...)
-	return nil
 }
